some big do-over for Emacs
Specifically:
- Switch to selectrum. It plays nice with existing Emacs facility, and feels
much faster than Ivy/Ido ever was.
- On that note, switch out smex and historian as well, to used a universal
prescient mode. The name was hard to remember though, but it has a lot of
integration which is nice.
- Switch back to Projectile. Why I'm I torturing myself with project.el? I mean
it is a good initiative but there is a long way for it to reach Projectile
maturity
- Added which-key, and add some pretty nice configuration for it
- No more Anzu, it's functions has been replaced (and improved by)
- Added visual-regex
- Ctrlf, experimenting with it now. It seems much faster than isearch,
noticeably less flashing and verbose at the minibuffer. Yanking to append to the
search query is missing, I'll look into it later